refactor: move SkeletonPostList to common component
