--- src/features/profile/components/SkeletonPostList.tsx
@@ -1,11 +1,0 @@
-import { PostSkeleton } from '~/features/posts/components'
-
-export const SkeletonPostList = (): JSX.Element => {
-  return (
-    <>
-      <PostSkeleton />
-      <PostSkeleton />
-      <PostSkeleton />
-    </>
-  )
-}
