+++ src/features/profile/components/index.ts
@@ -1,4 +1,3 @@
 export * from './ProfileDescription'
 export * from './ProfileTabs'
-export * from './SkeletonPostList'
 export * from './EmptyPostList'
